https://github.com/jordansissel/keynav/pull/49
Part of the X Nonrectangular Window Shape Extension includes the ability
to set the input shape. Setting the input shape to what is effectively a
0x0 rectangle results in a grid that cannot be clicked on; clicks and
motion events end up being passed to the window below the grid.

Since this basically achieves what {open,close}pixel() was originally
made for, we can just set the input shape as part of the initialization
for our window and remove {open,close}pixel() and mouseinfo, as they are
no longer needed.